Polish readme a bit
These are some minor edits I made while reading the readme.
Looking forward to trying out the tool tomorrow.

BTW, the readme declares jira-sync to be "A suite of utilities to synchronise jira projects to the local file system", while the GitHub project description reads "A utility to synchronise jira projects to the local file system". I was going to change the latter to match the former, but I don't seem to have write access to this repo.